# [PERF] Add a parallel local CSV reader