--- a/src/daft-csv/src/lib.rs
+++ b/src/daft-csv/src/lib.rs
@@ -1,5 +1,6 @@
 #![feature(async_closure)]
 #![feature(let_chains)]
+#![feature(new_uninit)]
 #![feature(trait_alias)]
 #![feature(trait_upcasting)]
 #![feature(test)]